Add GetDrivers feature to retrieve driver list
Introduced a new feature to retrieve a list of drivers from the database. Added the `GetDrivers` method to the business logic and repository layers, and updated the corresponding interfaces. Modified the `DriverController` to handle HTTP GET requests for drivers. Created a new `GetDriversResponse` DTO to structure the response data. Added a SQL query and resource entry to support this functionality.